界面结构调整

+++ src/main/java/yalong/site/json/entity/match/Player.java
@@ -17,7 +17,7 @@ public class Player {
     private String platformId;
     private String profileIcon;
     private String puuid;
-    private Integer summonerId;
+    private Long summonerId;
     private String summonerName;
     private String tagLine;
 }
